feat(provider-list): create component of provider list for service-account and collector
